Rename the component to prometheus.receive_http
